+++ src/vmware/azext_vmware/_help.py
@@ -127,6 +127,16 @@ helps['vmware private-cloud disable-cmk-encryption'] = """
     short-summary: Disable a Customer Managed Keys Encryption from a private cloud.
 """
 
+helps['vmware private-cloud add-extended-network-blocks'] = """
+    type: command
+    short-summary: Add extended network blocks to a private cloud.
+"""
+
+helps['vmware private-cloud delete-extended-network-blocks'] = """
+    type: command
+    short-summary: Delete extended network blocks from the private cloud.
+"""
+
 helps['vmware private-cloud identity'] = """
     type: group
     short-summary: Commands for Managed Identity in a private cloud.