+++ src/vmware/HISTORY.md
@@ -1,5 +1,10 @@
 # Release History
 
+## 6.0.1 (2023-09)
+- Update to AVS 2023-03-01 API
+- Add `az vmware private-cloud add-extended-network-blocks`
+- Add `az vmware private-cloud delete-extended-network-blocks`
+
 ## 5.0.3 (2023-07)
 
 - Fix the mandatory flag of `--hosts` for `az vmware cluster create` [#26653](https://github.com/Azure/azure-cli/issues/26653)